Default to float32, and allow float64

+++ mikeio/dfs1.py
@@ -13,7 +13,7 @@ from .spatial.grid_geometry import Grid1D
 class Dfs1(_Dfs123):
     _ndim = 1
 
-    def __init__(self, filename=None, dtype=np.float64):
+    def __init__(self, filename=None, dtype=np.float32):
         super(Dfs1, self).__init__(filename, dtype)
 
         self._dx = None
